Added. Validators for events APIs.

+++ app/pages/api/[id]/event/delete.ts
@@ -15,20 +15,34 @@ export default async function DeleteEvent(
         /** check user authorisation */
         await authorisedValidator(req, res);
 
-        /** deconstruct body data */
-        const { eventID } = req.body;
+        try {
+            /** deconstruct body data */
+            const { eventID } = req.body;
 
-        /* insert data into category table */
-        const result = await executeQuery({
-            query: 'DELETE FROM events WHERE AccountID=? AND ID=?',
-            values: [req.query.id, eventID],
-        });
+            if (eventID.length === 36) {
+                /* insert data into category table */
+                const result = await executeQuery({
+                    query: 'DELETE FROM events WHERE AccountID=? AND ID=?',
+                    values: [req.query.id, eventID],
+                });
 
-        res.status(200).json({ message: 'success' })
-        return
+                res.status(200).json({ message: 'success' })
+                return
+            }
+            /** if request body components do not fit requirements */
+            else {
+                res.statusCode = 400;
+                res.end('Request format error');
+            }
+        }
+        /** if request body components do not fit requirements */
+        catch {
+            res.statusCode = 400;
+            res.end('Request format error');
+        }
     }
-    /* rejects requests that are empty */
-    else if (!req.body) {
+    /* rejects requests that are not POST */
+    else if (req.method !== 'POST') {
         res.statusCode = 405;
         res.end('Error');
         return
